requirements.txt
numpy
pandas
pillow
streamlit
